Intergrated UserAuthenticationManager into ClientFacade
