+++ src/java/beans/ClientFacade.java
@@ -19,6 +19,8 @@ import Ebeans.ServiceForPAFacade;
 import Ebeans.ServiceProviderForPAFacade;
 import Ebeans.ServiceForPA;
 import Ebeans.ServiceProviderForPA;
+import Ebeans.UserData;
+import Ebeans.UserDataFacade;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -37,6 +39,12 @@ public class ClientFacade {
     @EJB
     private ReservationFacade reservationFacade;
     
+    @EJB
+    UserDataFacade userDataFacade;
+    
+    // For user authentication
+    UserData userData;
+    UserData userDataFound;
    
     //Produce Service Provider catalog wrap class to exchange between EJB server and Android client.
     public SPCatalogWrap produceSPcataWrap()
@@ -73,5 +81,79 @@ public class ClientFacade {
     {
         return reservationFacade.getReservations(uid);
     }
+  
+    /* userLogin
+     * Find ID in database and compare password
+     * return true or false
+     */
+    public boolean userLogin(UserData u) {
+        
+        
+        try {
+            userData = new UserData();
+            userData.setId(u.getId());
+            userData.setName(u.getName());
+            userData.setPwd(u.getPwd());
+            
+            userDataFound = new UserData();
+            userDataFound =  userDataFacade.find(u.getId());
+            
+            if(!userAlreadyExists(userData.getId())) {
+                return false;
+            }
+            return userData.getPwd().equals(userDataFound.getPwd());
+        } finally {
+        }
+    }
     
+    /* userRegister
+     * Create user data in database
+     */
+    public boolean userRegister(UserData u) {
+        UserData userData;
+        
+        try {
+            if(userAlreadyExists(u.getId()))
+                return false;
+            
+            userData = new UserData();
+            userData.setId(u.getId());
+            userData.setName(u.getName());
+            userData.setPwd(u.getPwd());
+                                    
+            userDataFacade.create(userData);
+            
+            System.out.println(userData.getId() + userData.getPwd() + userData.getName());
+            return true;
+        } finally {
+        }        
+    }
+    
+    /* userUnregister
+     * Remove user data from database
+     */
+    public boolean userUnregister(UserData u) {
+        UserData userData;
+        UserData userDataFound;
+        
+        try {
+            userData = new UserData();
+            userData.setId(u.getId());
+            userData.setName(u.getName());
+            userData.setPwd(u.getPwd());
+            
+            userDataFound = new UserData();            
+            userDataFound = userDataFacade.find(userData.getId());
+            if(userAlreadyExists(userData.getId()) && userData.getPwd().equals(userDataFound.getPwd())) {
+                userDataFacade.remove(userDataFound);
+                return true;
+            }
+            return false;
+        } finally {
+        }
+    }
+    
+    public boolean userAlreadyExists(String id) {
+        return userDataFacade.find(id)!=null;
+    }
 }
