+++ src/java/services/PamanagerWS.java
@@ -47,5 +47,4 @@ public class PamanagerWS {
     }
     
     
-    
 }
